+++ lib/date_value.dart
@@ -1,0 +1,6 @@
+class DateValue {
+  String startDate;
+  String endDate;
+
+  DateValue({this.startDate = "", this.endDate = ""});
+}