+++ lib/time_value.dart
@@ -1,0 +1,9 @@
+class TimeValue {
+  int seconds;
+  int minutes;
+  int hours;
+  int days;
+
+  TimeValue(
+      {this.seconds = 0, this.minutes = 0, this.hours = 0, this.days = 0});
+}